# openapi.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  version: 1.0.0
  description: REST API for managing projects, users, and tasks for the SymphonyAI Java Associate take-home assignment.
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /api/info:
    get:
      summary: Get application info
      responses:
        '200':
          description: Application information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
  /api/projects:
    post:
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      summary: List projects
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
      responses:
        '200':
          description: List or paged projects
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ProjectResponse'
                  - $ref: '#/components/schemas/PagedProjectResponse'
  /api/projects/{id}:
    get:
      summary: Get project by id
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update a project
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: Delete a project
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204': { description: Project deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/projects/{id}/tasks:
    get:
      summary: Get all tasks for a project
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: List of tasks for a project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '404': { $ref: '#/components/responses/NotFound' }
  /api/users:
    post:
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      summary: List users
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
      responses:
        '200':
          description: List or paged users
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  - $ref: '#/components/schemas/PagedUserResponse'
  /api/users/{id}:
    get:
      summary: Get user by id
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update a user
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: Delete a user
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204': { description: User deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/tasks:
    post:
      summary: Create a task under a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      summary: List tasks with optional filters and pagination
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
      responses:
        '200':
          description: List or paged tasks
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  - $ref: '#/components/schemas/PagedTaskResponse'
  /api/tasks/{id}:
    get:
      summary: Get task by id
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update task details or status
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete a task
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204': { description: Task deleted }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, example: 0 }
    SizeParam:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, example: 10 }
    SortByParam:
      name: sortBy
      in: query
      required: false
      schema: { type: string, example: id }
    SortDirParam:
      name: sortDir
      in: query
      required: false
      schema: { type: string, enum: [asc, desc], example: asc }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    InfoResponse:
      type: object
      properties:
        appName: { type: string, example: Task Tracker API }
        appVersion: { type: string, example: 1.0.0 }
    ProjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: Media Platform }
        description: { type: string, example: Task tracker backend }
    ProjectResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: Media Platform }
        description: { type: string, example: Task tracker backend }
        createdAt: { type: string, format: date-time, example: '2026-05-06T12:00:00' }
    UserRequest:
      type: object
      required: [name, email, role]
      properties:
        name: { type: string, example: Alice }
        email: { type: string, format: email, example: alice@example.com }
        role: { $ref: '#/components/schemas/UserRole' }
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: Alice }
        email: { type: string, example: alice@example.com }
        role: { type: string, example: DEVELOPER }
    TaskRequest:
      type: object
      required: [title, status, projectId]
      properties:
        title: { type: string, example: Implement project dashboard }
        status: { $ref: '#/components/schemas/TaskStatus' }
        priority: { $ref: '#/components/schemas/Priority' }
        projectId: { type: integer, format: int64, example: 1 }
        assigneeId: { type: integer, format: int64, nullable: true, example: 1 }
        dueDate: { type: string, format: date, example: '2026-05-20' }
    TaskResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        title: { type: string, example: Implement project dashboard }
        status: { type: string, example: IN_PROGRESS }
        priority: { type: string, example: HIGH }
        projectId: { type: integer, format: int64, example: 1 }
        projectName: { type: string, example: Media Platform }
        assigneeId: { type: integer, format: int64, nullable: true, example: 1 }
        assigneeName: { type: string, nullable: true, example: Alice }
        dueDate: { type: string, format: date, example: '2026-05-20' }
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 400 }
        error: { type: string, example: Bad Request }
        message: { type: string, example: Validation failed }
        path: { type: string, example: /api/users }
        details:
          type: array
          items: { type: string }
    PagedProjectResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/ProjectResponse' }
        page: { type: integer, example: 0 }
        size: { type: integer, example: 10 }
        totalElements: { type: integer, format: int64, example: 1 }
        totalPages: { type: integer, example: 1 }
        first: { type: boolean, example: true }
        last: { type: boolean, example: true }
    PagedUserResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/UserResponse' }
        page: { type: integer, example: 0 }
        size: { type: integer, example: 10 }
        totalElements: { type: integer, format: int64, example: 1 }
        totalPages: { type: integer, example: 1 }
        first: { type: boolean, example: true }
        last: { type: boolean, example: true }
    PagedTaskResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/TaskResponse' }
        page: { type: integer, example: 0 }
        size: { type: integer, example: 10 }
        totalElements: { type: integer, format: int64, example: 1 }
        totalPages: { type: integer, example: 1 }
        first: { type: boolean, example: true }
        last: { type: boolean, example: true }
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, DONE]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    UserRole:
      type: string
      enum: [DEVELOPER, QA, MANAGER]