# src/main/java/com/example/tasktracker/dto/response/UserResponse.java
package com.example.tasktracker.dto.response;

public class UserResponse {
    private Long id;
    private String name;
    private String email;
    private String role;

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
    public String getRole() { return role; }
    public void setRole(String role) { this.role = role; }
}